// src/app/shared/models/leads.ts
export class Leads {
    aadhaarCardNo: string
    annualIncome: string
    assignOn: Date
    assignToPartnerId: number
    assignedById: number
    cityId: number
    cityName: string
    contactNo: string
    courseId: number
    createdBy: number
    createdDate: Date
    customerFullName: string
    degreeId: number
    email: string
    employmentTypeId: number
    experience: string
    id: number
    isActive: boolean
    isDelete: boolean
    loanAmount: number
    loanPurpose: string
    modifiedBy: number
    modifiedDate: Date
    monthlyIncome: string
    professionType: string
    panCardNo: string
    pincode: string
    primaryBankId: number
    professionTypeId: number
    serviceId: number
    serviceName: string
    studiedCountryId: number
    bankName: string
    partnerName: string
    partnerContactNo: string
    statusHistory = [];
    createdByUser: string
    createdByContactNo: string
    customerId: number
    status: string
    label: string
    addressLine1: string
    addressLine2: string
    district: string
    state: string
    customerAddressId: number
    leadId: number
    constructor() {

    }
}